Google Calendar does not post duplicate notifications - use BasicMessageExtractor to avoid blocking recurrent reminders / events notifications
